Added gates for auth for update and delete

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -16,7 +16,11 @@ class AttendeeController extends Controller
      * Display a listing of the resource.
      */
 
-     use CanLoadRelationships;
+    use CanLoadRelationships;
+
+    public function __construct() {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
     private array $allowed = ['user'];
     public function index(Event $event)
     {
@@ -60,6 +64,7 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete-attendee',[$event, $attendee]);
         $attendee->delete();
 
         return response(status: 204);
